优化领域事件处理逻辑 修复LazyQueryCache未初始化的问题 修复DbContext在保存原值时空引用的问题 增加从DbContext.Local中获取entity的方法

+++ framework_modules/Geex.MongoDB.Entities/src/DB.Delete.cs
@@ -39,17 +39,21 @@ namespace MongoDB.Entities
             {
                 if (dbContext?.Local[typeof(T)].Remove(id, out var item) == true)
                 {
-                    foreach (var lazyQuery in item.LazyQueryCache.Values.Where(x => x.CascadeDelete))
+                    var lazyQeuryCacheValues = item.LazyQueryCache?.Values;
+                    if (lazyQeuryCacheValues?.Any() == true)
                     {
-                        var value = lazyQuery.Value;
-                        switch (value)
+                        foreach (var lazyQuery in lazyQeuryCacheValues.Where(x => x.CascadeDelete))
                         {
-                            case IQueryable<T> query:
-                                await query.DeleteAsync();
-                                break;
-                            case Lazy<T> lazy:
-                                await lazy.Value.DeleteAsync();
-                                break;
+                            var value = lazyQuery.Value;
+                            switch (value)
+                            {
+                                case IQueryable<T> query:
+                                    await query.DeleteAsync();
+                                    break;
+                                case Lazy<T> lazy:
+                                    await lazy.Value.DeleteAsync();
+                                    break;
+                            }
                         }
                     }
                 }
